[ALPS08982975] FPSGO: fix coverity
[Description]
fix below coverity issue:
ALPS08982975, ALPS08974268, ALPS08974156, ALPS08973886,
ALPS08973692, ALPS08973645, ALPS08973625, ALPS08973512,
ALPS08973492, ALPS08973466, ALPS08973446, ALPS08973430

[Test]
PDAS

MTK-Commit-Id: 88cce4db02d1daaea29dab481343c67f016a762c

CR-Id: ALPS08982975
Feature: FPSGO

+++ drivers/misc/mediatek/performance/fpsgo_v3/composer/src/fpsgo_adpf.c
@@ -21,7 +21,7 @@ static void fpsgo_do_adpf_hint(struct work_struct *work)
 
 	iter = container_of(work, struct fpsgo_adpf_session, sWork);
 
-	local_key = (local_key & iter->sid) | (iter->tgid << 8);
+	local_key = (local_key & iter->sid) | ((unsigned long long)iter->tgid << 8);
 	switch (iter->cmd) {
 	case ADPF_CREATE_HINT_SESSION:
 		fpsgo_ctrl2comp_user_create(iter->tgid, iter->tgid, local_key,
